requirements.txt
# ══════════════════════════════════════════════════════════════════════════════
# UC-03 Smart Triage — Python Requirements
# Install: pip install -r requirements.txt
# ══════════════════════════════════════════════════════════════════════════════

# ── System Prerequisite: Ollama (must be installed before running the app) ────
#
#   Linux / macOS:
#     curl -fsSL https://ollama.com/install.sh | sh
#
#   Windows:
#     Download installer from https://ollama.com/download
#
#   After installing Ollama, pull the model:
#     ollama pull llama3.1
#
#   Start the Ollama server (if not already running):
#     ollama serve
#
# ─────────────────────────────────────────────────────────────────────────────

# Excel file reading — used to load all UCM xlsx data files
openpyxl>=3.1.0
pandas>=2.0.0

# Environment variable management (.env file support)
python-dotenv>=1.0.0

# Streamlit — web UI
streamlit>=1.30.0
